pyproject.toml
[tool.poetry]
name = "tenant-test"
version = "0.1.0"
description = "Test application to test multi tenant usage in Django"
authors = ["Andréas Kühne <andreas@kuhne.se>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
django-tenants = "^3.6.1"
psycopg = {extras = ["binary"], version = "^3.1.19"}
django-tenant-users = "^1.5.0"
django-debug-toolbar = "^4.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
